requirements.txt
Flask
Flask-MySQLdb
mysqlclient
Werkzeug